+++ src/renderer/api/lyrics/lrcApi.ts
