+++ src/renderer/api/lyrics/lrcs/index.ts
