+++ src/renderer/api/lyrics/lrcs/karaoke/decrypt/inflate.ts
