+++ src/renderer/api/lyrics/lrcs/karaoke/decrypt/krc.ts
