+++ src/renderer/api/lyrics/lrcs/karaoke/decrypt/qrc.ts
