+++ src/renderer/api/lyrics/lrcs/karaoke/decrypt/qrcDes.ts
