+++ src/renderer/api/lyrics/lrcs/karaoke/index.ts
