+++ src/renderer/api/lyrics/lrcs/karaoke/parsers.ts
