+++ src/renderer/api/lyrics/lrcs/karaoke/providers/kugou.ts
