+++ src/renderer/api/lyrics/lrcs/karaoke/types.ts
