+++ src/renderer/api/lyrics/lrcs/kugou.ts
