+++ src/renderer/api/lyrics/lrcs/types.ts
