+++ src/renderer/components/states/maxExpand/components/setting/components/pluginMarket/WallpaperEditSection.tsx
@@ -33,7 +33,7 @@ import {
   normalizeWallpaperMarketListData,
   updateUserWallpaperMetadata,
   type WallpaperMarketItem,
-} from '../../../../../../../api/userAccountApi';
+} from '../../../../../../../api/user/userAccountApi';
 import { readLocalToken } from '../../../../../../../utils/userAccount';
 import { SvgIcon } from '../../../../../../../utils/SvgIcon';
 import { TagInput } from './TagInput';
