+++ src/renderer/components/states/sliderCaptcha/SliderCaptchaContent.tsx
@@ -23,7 +23,7 @@
 import { useMemo, useState } from 'react';
 import type { CSSProperties, ReactElement } from 'react';
 import { useTranslation } from 'react-i18next';
-import type { UserCaptchaChallenge } from '../../../api/userAccountApi';
+import type { UserCaptchaChallenge } from '../../../api/user/userAccountApi';
 import '../../../styles/slider-captcha.css';
 import eislandLogo from '../../../../../resources/icon/eisland.svg';
