+++ src/renderer/utils/sliderCaptcha.ts
@@ -24,7 +24,7 @@
  * @author 鸡哥
  */
 
-import { createUserCaptchaChallenge, fetchUserCaptchaConfig, type UserCaptchaChallenge } from '../api/userAccountApi';
+import { createUserCaptchaChallenge, fetchUserCaptchaConfig, type UserCaptchaChallenge } from '../api/user/userAccountApi';
 import { createElement } from 'react';
 import { createRoot } from 'react-dom/client';
 import { SliderCaptchaContent } from '../components/states/sliderCaptcha/SliderCaptchaContent';
